component & file rename
